Become compatible with React Native v0.59
React Native v0.59 uses a new version of the Android SDK that is incompatible with the one hard-coded in this library, producing build failures. Using the rootProject setting means this library will automatically work with a broad range of React Native versions, so this will improve compatibility with a broad range of React Native versions.